Add login functionality and error handling to LoginScreen
This commit adds the login functionality to the LoginScreen component. It includes the implementation of the handleLogin function, which sends a POST request to the "/auth/login" endpoint using the AuthService. If the login is successful, the onLoginSuccess function from the AuthContext is called to update the authentication state. If there is an error, an alert is displayed with the error message.